update ui src dir path for workflows with incorrect path (#301)
## Summary

Most of the workflows had the incorrect path for where to pull in the
src dir of the build that is to be added to the gh-pages branch. It
resulted in the build silently not doing anything.

## Test Plan

- I'd like for there to be a way to test this, but ultimately I just
need to wait and see that the builds show up in the gh-pages branch when
things are merged into main and when the nightly runs. I think it is
virtually a guarantee it will be fixed though.

Signed-off-by: dalthecow <dalcowboiz@gmail.com>